stage4/analytics-dashboard/requirements.txt
web3>=6.0.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
